[Refactor] Consolidate DCP attention operations
