+++ vllm/v1/attention/ops/common.py
@@ -2,306 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 import torch
 
-from vllm.distributed.parallel_state import GroupCoordinator
 from vllm.triton_utils import tl, triton
-
-
-def mask_dcp_empty_shards_(
-    lse: torch.Tensor,
-    seq_lens: torch.Tensor | None,
-    query_start_loc: torch.Tensor | None,
-) -> None:
-    if seq_lens is None and query_start_loc is None:
-        return
-    if seq_lens is None or query_start_loc is None:
-        raise ValueError("seq_lens and query_start_loc must be provided together")
-    if (
-        seq_lens.ndim != 1
-        or query_start_loc.ndim != 1
-        or query_start_loc.shape[0] != seq_lens.shape[0] + 1
-    ):
-        raise ValueError("query_start_loc must contain one boundary per sequence")
-
-    row_indices = torch.arange(
-        lse.shape[0], device=lse.device, dtype=query_start_loc.dtype
-    )
-    sequence_indices = torch.searchsorted(
-        query_start_loc[1:], row_indices, right=True
-    ).clamp_max(seq_lens.shape[0] - 1)
-    empty_rows = (row_indices >= query_start_loc[-1]) | (
-        seq_lens[sequence_indices] == 0
-    )
-    lse.masked_fill_(empty_rows[:, None], float("-inf"))
-
-
-@triton.jit
-def _correct_attn_cp_out_kernel(
-    outputs_ptr,
-    new_output_ptr,
-    lses_ptr,
-    vlse_ptr,
-    outputs_stride_B,
-    outputs_stride_H,
-    outputs_stride_D,
-    lses_stride_N,
-    lses_stride_B,
-    lses_stride_H,
-    lse_idx,
-    HEAD_DIM: tl.constexpr,
-    N_ROUNDED: tl.constexpr,
-    IS_BASE_E: tl.constexpr,
-):
-    """
-    Apply the all-gathered lses to correct each local rank's attention
-    output. we still need perform a cross-rank reduction to obtain the
-    final attention output.
-
-    Args:
-        outputs_ptr (triton.PointerType):
-            Pointer to input tensor of shape [ B, H, D ]
-        lses_ptr (triton.PointerType):
-            Pointer to input tensor of shape [ N, B, H ]
-        new_output_ptr (triton.PointerType):
-            Pointer to output tensor of shape [ B, H, D ]
-        vlse_ptr (triton.PointerType):
-            Pointer to output tensor of shape [ B, H ]
-    """
-    batch_idx = tl.program_id(axis=0).to(tl.int64)
-    head_idx = tl.program_id(axis=1).to(tl.int64)
-    d_offsets = tl.arange(0, HEAD_DIM)
-    num_n_offsets = tl.arange(0, N_ROUNDED)
-
-    # shape = [N]
-    lse_offsets = (
-        num_n_offsets * lses_stride_N
-        + batch_idx * lses_stride_B
-        + head_idx * lses_stride_H
-    )
-
-    # calc final lse
-    lse = tl.load(lses_ptr + lse_offsets).to(tl.float32)
-    lse = tl.where((lse != lse) | (lse == float("inf")), -float("inf"), lse)
-    lse_max = tl.max(lse, axis=0)
-    lse_max = tl.where(lse_max == -float("inf"), 0, lse_max)
-    lse -= lse_max
-    if IS_BASE_E:
-        lse_exp = tl.exp(lse)
-        lse_acc = tl.sum(lse_exp, axis=0)
-        lse = tl.log(lse_acc)
-    else:
-        lse_exp = tl.exp2(lse)
-        lse_acc = tl.sum(lse_exp, axis=0)
-        lse = tl.log2(lse_acc)
-    lse += lse_max
-
-    lse_offsets = batch_idx * lses_stride_B + head_idx * lses_stride_H
-    tl.store(vlse_ptr + lse_offsets, lse)
-
-    # shape = [D]
-    output_offsets = (
-        batch_idx * outputs_stride_B
-        + head_idx * outputs_stride_H
-        + d_offsets * outputs_stride_D
-    )
-
-    # correct output
-    lse_offset = (
-        lse_idx * lses_stride_N + batch_idx * lses_stride_B + head_idx * lses_stride_H
-    )
-    lse_tmp = tl.load(lses_ptr + lse_offset).to(tl.float32)
-    lse_finally = lse_tmp - lse
-    lse_finally = tl.where(
-        (lse_finally != lse_finally) | (lse_finally == float("inf")),
-        -float("inf"),
-        lse_finally,
-    )
-    factor = tl.exp(lse_finally) if IS_BASE_E else tl.exp2(lse_finally)
-    output = tl.load(outputs_ptr + output_offsets)
-    output = output * factor
-    output = tl.where(factor == 0.0, 0.0, output)
-
-    tl.store(new_output_ptr + output_offsets, output)
-
-
-class CPTritonContext:
-    """The CPTritonContext is used to avoid recompilation of the Triton JIT."""
-
-    def __init__(self):
-        self.inner_kernel = None
-
-    def call_kernel(self, kernel, grid, *regular_args, **const_args):
-        if self.inner_kernel is None:
-            self.inner_kernel = kernel[grid](*regular_args, **const_args)
-        else:
-            self.inner_kernel[grid](*regular_args)
-
-
-def correct_attn_out(
-    out: torch.Tensor,
-    lses: torch.Tensor,
-    cp_rank: int,
-    ctx: CPTritonContext,
-    is_lse_base_on_e: bool = True,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Correct the attention output using the all-gathered lses.
-
-    Args:
-        out: Tensor of shape [ B, H, D ]
-        lses: Tensor of shape [ N, B, H ]
-        cp_rank: Current rank in the context-parallel group
-        ctx: Triton context to avoid recompilation
-
-    Returns:
-        Tuple of (out, lse) with corrected attention and final log-sum-exp.
-    """
-    if ctx is None:
-        ctx = CPTritonContext()
-
-    # --- Normalize to 3D views ---
-    if out.ndim == 4 and out.shape[1] == 1:
-        out = out.squeeze(1)
-    assert out.ndim == 3, f"expected out [B,H,D] or [B,1,H,D], got {tuple(out.shape)}"
-
-    if lses.ndim == 4 and lses.shape[-1] == 1:
-        lses = lses.squeeze(-1)
-    if lses.ndim == 4 and lses.shape[1] == 1:
-        lses = lses.squeeze(1)
-    assert lses.ndim == 3, (
-        f"expected lses [N,B,H] (optionally with a 1-sized extra dim), "
-        f"got {tuple(lses.shape)}"
-    )
-
-    B, H, D = out.shape
-    N = lses.shape[0]
-
-    # Strides after we normalized shapes to 3-D views.  The kernel computes
-    # offsets for `vlse_ptr` using lses_stride_B/H, so the output buffer must
-    # have the same B/H stride layout as a slice of `lses`.
-    o_sB, o_sH, o_sD = out.stride()
-    l_sN, l_sB, l_sH = lses.stride()
-
-    # Allocate LSE with the same B/H strides as `lses` so writes land correctly
-    # even when `lses` is a non-contiguous view (e.g., 4-D to 3-D squeeze).
-    lse = torch.empty_strided(
-        (B, H), (l_sB, l_sH), device=lses.device, dtype=lses.dtype
-    )
-
-    # Kernel launch config
-    grid = (B, H, 1)
-
-    regular_args = (
-        out,
-        out,
-        lses,
-        lse,
-        o_sB,
-        o_sH,
-        o_sD,
-        l_sN,
-        l_sB,
-        l_sH,
-        cp_rank,
-    )
-    const_args = {"HEAD_DIM": D, "N_ROUNDED": N, "IS_BASE_E": is_lse_base_on_e}
-    ctx.call_kernel(_correct_attn_cp_out_kernel, grid, *regular_args, **const_args)
-    return out, lse
-
-
-def _cp_lse_common(
-    cp_attn_out: torch.Tensor,
-    cp_attn_lse: torch.Tensor,
-    cp_group: GroupCoordinator,
-    ctx: CPTritonContext | None = None,
-    is_lse_base_on_e=True,
-    seq_lens: torch.Tensor | None = None,
-    query_start_loc: torch.Tensor | None = None,
-):
-    """
-    cp_attn_out: [ B, H, D ]
-    cp_attn_lse: [ B, H ]
-    """
-    if cp_group.world_size == 1:
-        return cp_attn_out
-
-    if ctx is None:
-        ctx = CPTritonContext()
-
-    cp_attn_lse = cp_attn_lse.contiguous()
-    mask_dcp_empty_shards_(cp_attn_lse, seq_lens, query_start_loc)
-    lses = cp_group.all_gather(cp_attn_lse, dim=0).reshape(
-        (cp_group.world_size,) + cp_attn_lse.shape
-    )
-    out, lse = correct_attn_out(
-        cp_attn_out,
-        lses,
-        cp_group.rank_in_group,
-        ctx,
-        is_lse_base_on_e=is_lse_base_on_e,
-    )
-    return out, lse
-
-
-def cp_lse_ag_out_rs(
-    cp_attn_out: torch.Tensor,
-    cp_attn_lse: torch.Tensor,
-    cp_group: GroupCoordinator,
-    ctx: CPTritonContext | None = None,
-    return_lse: bool = False,
-    is_lse_base_on_e=True,
-    seq_lens: torch.Tensor | None = None,
-    query_start_loc: torch.Tensor | None = None,
-):
-    """
-    cp_attn_out: [ B, H, D ]
-    cp_attn_lse: [ B, H ]
-    """
-    out, lse = _cp_lse_common(
-        cp_attn_out,
-        cp_attn_lse,
-        cp_group,
-        ctx=ctx,
-        is_lse_base_on_e=is_lse_base_on_e,
-        seq_lens=seq_lens,
-        query_start_loc=query_start_loc,
-    )
-    out = cp_group.reduce_scatter(out, dim=1)
-
-    if return_lse:
-        cp_num_heads = lse.shape[1] // cp_group.world_size
-        cp_rank = cp_group.rank_in_group
-        lse = lse[:, cp_num_heads * cp_rank : cp_num_heads * (cp_rank + 1)]
-        return out, lse
-    return out
-
-
-def cp_lse_ag_out_ar(
-    cp_attn_out: torch.Tensor,
-    cp_attn_lse: torch.Tensor,
-    cp_group: GroupCoordinator,
-    ctx: CPTritonContext | None = None,
-    return_lse: bool = False,
-    is_lse_base_on_e=True,
-    seq_lens: torch.Tensor | None = None,
-    query_start_loc: torch.Tensor | None = None,
-):
-    """
-    cp_attn_out: [ B, H, D ]
-    cp_attn_lse: [ B, H ]
-    """
-    out, lse = _cp_lse_common(
-        cp_attn_out,
-        cp_attn_lse,
-        cp_group,
-        ctx=ctx,
-        is_lse_base_on_e=is_lse_base_on_e,
-        seq_lens=seq_lens,
-        query_start_loc=query_start_loc,
-    )
-    out = cp_group.all_reduce(out)
-
-    if return_lse:
-        return out, lse
-    return out
 
 
 @triton.jit
